chore: reorganize data files
